test: migrate to Jest, add documentation for text extraction tests
